Distribute and remove input-valdation.test.ts
