+++ source/patterns/@aws-solutions-constructs/core/test/cloudfront-distribution-s3-helper.test.ts
@@ -21,6 +21,7 @@ import { Bucket } from 'aws-cdk-lib/aws-s3';
 import * as origins from 'aws-cdk-lib/aws-cloudfront-origins';
 import { LambdaEdgeEventType } from 'aws-cdk-lib/aws-cloudfront';
 import * as acm from 'aws-cdk-lib/aws-certificatemanager';
+import * as defaults from '../';
 
 test('check bucket policy metadata', () => {
   const stack = new Stack();
@@ -538,4 +539,24 @@ test('test cloudfront override cloudfront custom domain names ', () => {
       ],
     }
   });
+});
+
+// ---------------------------
+// Prop Tests
+// ---------------------------
+test('Test CloudFront insertHttpHeaders bad props', () => {
+
+  const props: defaults.CloudFrontProps = {
+    insertHttpSecurityHeaders: true,
+    responseHeadersPolicyProps: {
+      securityHeadersBehavior: {}
+    }
+  };
+
+  const app = () => {
+    defaults.CheckCloudFrontProps(props);
+  };
+
+  // Assertion
+  expect(app).toThrowError('responseHeadersPolicyProps.securityHeadersBehavior can only be passed if httpSecurityHeaders is set to `false`.');
 });
